# lab3/src/main/resources/API.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lab 2 Rest API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/company/employees/{id}": {
      "get": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Получить данные сотрудника",
        "description": "Получение данных сотрудника по его ID",
        "operationId": "read_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Добавить/изменить сотрудника",
        "description": "Добавление/изменение сотрудника в зависимости от существования указанного ID",
        "operationId": "create_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "ФИО",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "department",
            "in": "query",
            "description": "Отдел",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "position",
            "in": "query",
            "description": "Должность",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Удалить сотрудника",
        "description": "Удаление сотрудника по ID",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/company/access/{e_id}-{r_id}": {
      "get": {
        "tags": [
          "access-rest-controller"
        ],
        "summary": "Проверить доступ сотрудника",
        "description": "Проверка доступа сотрудника к помещению",
        "operationId": "read_5",
        "parameters": [
          {
            "name": "e_id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "r_id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "access-rest-controller"
        ],
        "summary": "Выдать доступ сотруднику",
        "description": "Для выдачи доступа сотрудник и помещение должны существовать, иначе - 404",
        "operationId": "create_2",
        "parameters": [
          {
            "name": "e_id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "r_id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccessDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "access-rest-controller"
        ],
        "summary": "Отозвать доступ у сотрудника",
        "description": "Отзыв доступа к помещению у сотрудника",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "e_id",
            "in": "path",
            "description": "ID сотрудника",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "r_id",
            "in": "path",
            "description": "ID помещения",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/company/employees": {
      "get": {
        "tags": [
          "employee-rest-controller"
        ],
        "summary": "Получить список сотрудников",
        "description": "Получение всех сотрудников и их данных",
        "operationId": "read_2",
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/company/access": {
      "get": {
        "tags": [
          "access-rest-controller"
        ],
        "summary": "Получить список доступов",
        "description": "Получение всех выданных сотрудникам доступов",
        "operationId": "read_4",
        "responses": {
          "200": {
            "description": "Успешное выполнение",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessDTO"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID сотрудника",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Имя сотрудника"
          },
          "department": {
            "type": "string",
            "description": "Отдел, где работает сотрудник"
          },
          "position": {
            "type": "string",
            "description": "Должность сотрудника"
          }
        },
        "description": "Сотрудник"
      },
      "AccessDTO": {
        "type": "object",
        "properties": {
          "employeeID": {
            "type": "integer",
            "description": "ID сотрудника",
            "format": "int32"
          },
          "roomID": {
            "type": "integer",
            "description": "ID помещения",
            "format": "int32"
          }
        },
        "description": "Доступ к помещению"
      }
    }
  }
}